requirements.txt
async-timeout==5.0.1
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.3.2
convertdate==2.4.0
cycler==0.12.1
dash==3.0.4
dash-iconify==0.1.2
dash_mantine_components==2.0.0
Django==5.1.6
django-admin-rangefilter==0.13.2
django-autocomplete-light==3.12.1
django-countries==7.6.1
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-json-widget==2.0.1
django-more-admin-filters==1.12
django-mptt==0.17.0
django-mptt-admin==2.8.0
django-treebeard==4.7.1
et_xmlfile==2.0.0
Flask==3.0.3
fonttools==4.58.2
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
lunardate==0.2.2
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.3
narwhals==1.37.1
nest-asyncio==1.6.0
numpy==2.2.3
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pillow==11.1.0
plotly==6.0.1
pyluach==2.2.0
PyMeeus==0.5.12
PyMySQL==1.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.1
redis==5.2.1
requests==2.32.3
retrying==1.3.4
setuptools==80.1.0
six==1.17.0
SQLAlchemy==2.0.38
sqlparse==0.5.3
tabulate==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.4.0
Werkzeug==3.0.6
zipp==3.21.0
asgiref>=3.8.1
async-timeout==5.0.1
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.3.2
convertdate==2.4.0
cycler==0.12.1
dash==3.0.4
dash-iconify==0.1.2
dash_ag_grid==31.3.1
dash_mantine_components==2.0.0
Django==5.1.6
django-admin-rangefilter==0.13.2
django-autocomplete-light==3.12.1
django-countries==7.6.1
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-json-widget==2.0.1
django-more-admin-filters==1.12
django-mptt==0.17.0
django-mptt-admin==2.8.0
django-treebeard==4.7.1
dotenv==0.9.9
et_xmlfile==2.0.0
Flask==3.0.3
fonttools==4.58.2
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
lunardate==0.2.2
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.3
narwhals==1.37.1
nest-asyncio==1.6.0
numpy==2.2.3
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pillow==11.1.0
plotly==6.0.1
pyluach==2.2.0
PyMeeus==0.5.12
PyMySQL==1.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.1
redis==5.2.1
requests==2.32.3
retrying==1.3.4
setuptools==80.1.0
six==1.17.0
SQLAlchemy==2.0.38
sqlparse==0.5.3
tabulate==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.4.0
Werkzeug==3.0.6
zipp==3.21.0
async-timeout==5.0.1
blinker==1.9.0
certifi==2025.4.26
channels==4.3.2
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.3.2
convertdate==2.4.0
cycler==0.12.1
dash==3.0.4
dash-bootstrap-components==2.0.4
dash-iconify==0.1.2
dash_ag_grid==31.3.1
dash_mantine_components==2.0.0
Django==5.1.6
django-admin-rangefilter==0.13.2
django-autocomplete-light==3.12.1
django-countries==7.6.1
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-json-widget==2.0.1
django-more-admin-filters==1.12
django-mptt==0.17.0
django-mptt-admin==2.8.0
django-plotly-dash==2.5.0
django-treebeard==4.7.1
dotenv==0.9.9
dpd_components==0.2.0
et_xmlfile==2.0.0
Flask==3.0.3
fonttools==4.58.2
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
lunardate==0.2.2
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.3
narwhals==1.37.1
nest-asyncio==1.6.0
numpy==2.2.3
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pillow==11.1.0
plotly==6.0.1
pyluach==2.2.0
PyMeeus==0.5.12
PyMySQL==1.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.1
redis==5.2.1
requests==2.32.3
retrying==1.3.4
setuptools==80.1.0
six==1.17.0
SQLAlchemy==2.0.38
sqlparse==0.5.3
tabulate==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.4.0
Werkzeug==3.0.6
zipp==3.21.0
